pyproject.toml
[tool.poetry]
name = "strana-bot"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tqdm = "^4.66.4"
chromadb = "^0.5.3"
openai = "^1.35.1"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.17"
loguru = "^0.7.2"
bs4 = "^0.0.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
aiogram = "^3.8.0"
ipywidgets = "^8.1.3"
redis = "^5.0.6"
oauth2client = "^4.1.3"
tiktoken = "^0.7.0"
gspread = "^6.1.2"
openpyxl = "^3.1.4"
speechrecognition = "^3.10.4"
psycopg2-binary = "^2.9.9"
update = "^0.0.1"
aiohttp = "^3.9.5"
instabot = "^0.117.0"
schedule = "^1.2.2"
langchain-openai = "^0.1.23"
gigachain-community = "^0.2.16.post3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
